+++ R/environments.R
@@ -16,14 +16,18 @@
 #' * `environments()`: Invisibly, a `character` vector of environment names
 #' @name list_environments
 environments <- function() {
-  s <- search()
+  x <- search()
+  class(x) <- c("character", "mark_environments")
+  x
+}
 
-  for (e in s) {
-    # TODO update to a print.*() method
+#' @exportS3Method
+#' @rdname list_environments
+print.mark_environments <- function(x, ...) {
+  for (e in x) {
     cat(utils::str(parent.env(as.environment(e)), give.attr = FALSE))
   }
-
-  invisible(s)
+  invisible(x)
 }
 
 #' @export